# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Audio API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tags": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Вывод списка тэгов с постами",
                "description": "Список тэгов с пагинацией",
                "operationId": "65f85cc85cd13d2454b6c88ffff5cc3a",
                "responses": {
                    "200": {
                        "description": "Список",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tag"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/tags/{tag}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Вывод одного тэга с постами",
                "operationId": "890706647a0186ae7289ff232e056564",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID тэга",
                            "type": "integer",
                            "example": "5"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Вывод поста",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tag"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Tag": {
                "description": "Тэги и посты",
                "required": [
                    "id",
                    "name",
                    "posts"
                ],
                "properties": {
                    "title": {
                        "description": "Загаловок тэга",
                        "type": "string",
                        "example": "Заголовок"
                    },
                    "posts": {
                        "description": "Посты с этим тэгом",
                        "type": "array",
                        "items": {
                            "properties": {
                                "title": {
                                    "description": "Загаловок поста",
                                    "type": "string",
                                    "example": "Заголовок"
                                },
                                "description": {
                                    "description": "Описание поста",
                                    "type": "text",
                                    "example": "Описание поста с аудио-файлом"
                                },
                                "user_id": {
                                    "description": "ID автора поста",
                                    "type": "integer",
                                    "example": 4
                                },
                                "is_active": {
                                    "description": "Статус поста",
                                    "type": "boolean",
                                    "example": 1
                                },
                                "file_path": {
                                    "type": "string",
                                    "example": "compressed/audio_name.mp3"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}